Stop trying trying to transfer after failure
When a device is detached, the next IN or OUT transfer will result
in XFER_RESULT_FAILED. On Pico-PIO-USB host devices connected via a
hub, this can result in the next poll of hub status to fail. It is
important to not attempt to retry transfers to a device that is
not attached or else the hub will never be able to report status
properly and report to the system that the device is detached.